fix: SurroundModesMessenger uses base device key
Because of the way the surround modes messenger was built,
it was using the key of the parent device in the path, eg /device/avr,
but the key of the ISelectableItems object in the actual messages. This
caused the key of the device to change in the store, which caused cascading issues
in the UI, as the device key was expected to remain as the parent device's key.